move: query default parser main functions to utils
